tambahan dan perbaikan

+++ app/src/main/java/com/athalah/valuta/data/Currency.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 
 data class Currency (
     val name: String,
-    val buy: Float,
-    val sell: Float,
+    val buy: Double,
+    val sell: Double,
     val flag: String
 )
